To push server updates if a new webSocket listener is added/set
By default it's enabled, it can be disabled/enabled by
BrowserPage#setPushQueueOnNewWebSocketListener method.

Example use case : client clicked on a button to update UI. The button
click event is received at server side and the changes took place at
server side. But, when the server tried to push the UI updates to the
client the webSocket connection lost so the UI updates are kept in
queue. After that, the client made a successful webSocket connection but
the server (UI) updates will not be pushed until there is something to
push again. The client clicked on the button again but as there is
nothing to change in the server there won't be anything to push from the
server. To fix this minor bug, it's required to push server updates
queue when new webSocket listener is added/set.