Implemented feature for storing some heap memory data to external drive
  -- Implemented feature to save temporary heap memory data (i.e. kept
in queue for pushing and incoming data cached in queue) to an external
storage drive. This will prevent out of memory issue up to a certain
level, for instance if the heap size allocated for the application is
not enough with respect to the number of threads running in the app or
there is a heavy traffic due to action events from the end users browser
pages. BrowserPage.useExternalDrivePath method may be overridden to set
the path for such temporary files.
  -- Test case is also added.